backend/requirements.txt
# -------------------- Django Core Stack -------------------- #
Django>=4.2
djangorestframework>=3.14  # REST API Framework
psycopg2-binary>=2.9  # PostgreSQL Adapter
django-cors-headers>=4.3  # CORS Handling
django-environ>=0.11.2  # Environment Variable Management
djangorestframework-simplejwt>=5.3  # JWT Authentication
PyJWT>=2.8  # JSON Web Token Support

# -------------------- Google Gemini AI SDK -------------------- #
# google-generativeai>=0.5.0
google-api-core>=2.17.1
langchain>=0.1.14  # Langchain for LLM Pipeline
google-ai-generativelanguage==0.6.15
# -------------------- Data Parsing & Processing -------------------- #
pandas>=2.1
openpyxl>=3.1  # Excel File Handling
pdfplumber>=0.10  # PDF Parsing
PyMuPDF>=1.23.7  # PDF Text Extraction (if needed)
pdfminer.six>=20221105  # Alternative PDF Parsing
cloudinary>=1.43.0  # Cloudinary Image Uploads

# -------------------- LLM & Embedding-based Search -------------------- #
# transformers>=4.38  # Hugging Face Model Support (Optional for RAG)
# torch>=2.2  # Required for Deep Learning
# sentence-transformers>=2.2  # For Embedding-based SDG Mapping
langchain-openai>=0.0.5  # Future integration with OpenAI models (optional)
# pgvector>=0.2.5  # Store Embeddings in PostgreSQL

# -------------------- Async & Caching -------------------- #
celery>=5.3  # Background Task Handling
redis>=5.0  # Redis Caching Layer

# -------------------- Environment & Security -------------------- #
python-dotenv>=1.0.1  # Load .env Files
requests>=2.31  # API Requests Handling
httpx>=0.25  # For Async HTTP Handling

# -------------------- Frontend Dependencies (Next.js) -------------------- #
# Add in `package.json` (if needed)
